Fixed some warnings from recently added tests.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@6203 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rwlibs/algorithms/RANSACModel.hpp
@@ -127,7 +127,7 @@ public:
 			// if consensus set size is large enough, we have a model
 			if (consensusSet.size() > static_cast<size_t>(dataRequired)) {
 				models.push_back(
-						std::pair<MODEL, int>(maybeModel, consensusSet.size()));
+						std::pair<MODEL, int>(maybeModel, static_cast<int>(consensusSet.size())));
 			}
 
 		}
